created separate components for header, form and accounts list. Closes #1
