+++ src/assets/style/_component.account-form.scss
@@ -68,6 +68,10 @@
     &--sm-margin {
       margin-bottom: 29px;
     }
+
+    &--gender {
+      margin-bottom: 26px;
+    }
   }
 
   &__field-container {
