+++ src/mixins.js
@@ -1,0 +1,13 @@
+// eslint-disable-next-line import/prefer-default-export
+export const validations = {
+  getAge(birthday) {
+    const today = new Date();
+    const birthDate = new Date(birthday);
+    let age = today.getFullYear() - birthDate.getFullYear();
+    const m = today.getMonth() - birthDate.getMonth();
+    if (m < 0 || (m === 0 && today.getDate() < birthDate.getDate())) {
+      age -= 1;
+    }
+    return age;
+  },
+};
